examples-new/pitchdeck_eval_code_entry/requirements.txt
# Requirements for pitch deck evaluation code entry example
python-slugify>=8.0
